[BUGFIX] Preventing object storage from saving empty values
Unlike normal data from form submissions, values for files will be submitted whether a file was uploaded or not. So when other data doesn't get saved to an ObjectStorage object, file uploads do. And, of course, we don't want that. But, really, why would we want an ObjectStorage to hold empty values in the first place? So, the best way to deal with this is to prevent empty values from being saved to an ObjectStorage object.

Yes, this is really only for file uploads. But the principle works globally.