[F] Add 'forgiving' feature for migrations
Because there are alternate paths to database updates being made,
sometimes it makes sense to skip certain migrations if they are not
necessary, or would cause errors. For example, if a migration adds
a field to a table, you don't want it to fail when/if the field
already exists (via a content sync mechanism, etc.). We want to be
able to run migrations with whatever build/deploy mechanism we
are using, and a "stuck" migration would make successful deployment
impossible.