[C] Disable staticImage fileExists test for now
This test doesn't work due to conflicts with the internal caching in the
StaticImageService class. Need to mock the caching part of it, maybe
(but it's hard to get Extbase instantiation at this point, using mocks).
Cheaper/flimsier alternative is to allow for a "ignore cache" option on
the interface, maybe...